+++ Blurred-ios/Settings/AccountSettingsViewController.swift
@@ -47,8 +47,8 @@ class AccountSettingsViewController: UIViewController {
         DispatchQueue.main.async {
             self.view.addSubview(myActivityIndicator)
         }
-        let userId: String? = myValet.string(forKey: "Id")
-        let token: String? = tokenValet.string(forKey: "Token")
+        let userId: String? = try? myValet.string(forKey: "Id")
+        let token: String? = try? tokenValet.string(forKey: "Token")
         let Id = Int(userId!)
         let myUrl = URL(string: "http://10.0.0.2:3000/api/v1/registrations/\(Id!).json")
         var request = URLRequest(url:myUrl!)